refactor: 예매 내역 리스트 데이터 바인딩으로 변경
